Got tilemap, bomb, collisions, etc working.

+++ source/PlayState.hx
@@ -5,6 +5,7 @@ import flixel.FlxSprite;
 import flixel.FlxState;
 import flixel.text.FlxText;
 import flixel.ui.FlxButton;
+import flixel.group.FlxTypedGroup;
 import flixel.util.FlxMath;
 
 enum PlayMode {
@@ -22,7 +23,11 @@ class PlayState extends FlxState
 	var testLevel:WorldLevel;
 	var player:Player;
 
+	public var bombs = new FlxTypedGroup<Bomb>();
+
 	public var activeLevel:WorldLevel;
+
+	var loadingBar:LoadingBar;
 
 	override public function create():Void
 	{
@@ -31,7 +36,12 @@ class PlayState extends FlxState
 		testLevel = new WorldLevel("assets/levels/TestLevel.tmx");
 		add(testLevel.allTilemaps);
 
+		loadingBar =  new LoadingBar(0,0,this);
+		add(loadingBar);
+
 		activeLevel = testLevel;
+
+		add(bombs);
 
 
 		FlxG.camera.follow(player);
@@ -46,14 +56,37 @@ class PlayState extends FlxState
 	}
 
 	function startRunMode () {
+
 		player = new Player (40,40,this);
+		loadingBar.setProgressValue(0);
 		FlxG.camera.follow(player,0);
 		add(player);
+
+	}
+
+	function updateRunMode () {
+
+	}
+	function updateShootMode () {
+		loadingBar.load((FlxG.elapsed )/5);
+
+		if (FlxG.keys.justPressed.SPACE || FlxG.mouse.justPressed){
+			var bomb = new Bomb(player.x,player.y,this);
+			bomb.velocity.x = FlxG.mouse.x - player.x;
+			bomb.velocity.y = FlxG.mouse.y - player.y;
+			bombs.add(bomb);
+		}
+
+		if (loadingBar.progress > 1){
+			player.kill();
+			mode = PlayMode.running;
+			resetWorld();
+		}
 	}
 
 	function startShootingMode () {
 		//Argh tired need to sleep lol.
-		player.kill();
+		player.active = false;
 
 	}
 
@@ -72,12 +105,11 @@ class PlayState extends FlxState
 	override public function update():Void
 	{
 		super.update();
-
+		//Awful code. Not for educational research. Ever.
 		if (mode == PlayMode.shooting) {
-			if (FlxG.keys.justPressed.SPACE){
-				mode = PlayMode.running;
-				resetWorld();
-			}
+			updateShootMode();
+		}else{
+			updateRunMode();
 		}
 
 	}
